Allow admins to unlink lti contexts
Also fixed alembic migrations so they use static enum classes for migrations instead of current versions and unique constraint warning will no longer appear in tests.
Added uuids to third_party_user and lti_user tables in preparation for #517

Closes #579
