+++ compair/api/lti_launch.py
@@ -26,11 +26,6 @@ api = new_restful_api(lti_api)
 lti_launch_parser = RequestParser()
 lti_launch_parser.add_argument('assignment', default=None)
 
-# events
-on_lti_course_link = event.signal('LTI_CONTEXT_COURSE_LINKED')
-on_lti_course_membership_update = event.signal('LTI_CONTEXT_COURSE_MEMBERSHIP_UPDATE')
-on_lti_course_membership_status_get = event.signal('LTI_CONTEXT_COURSE_MEMBERSHIP_STATUS_GET')
-
 # /auth
 class LTIAuthAPI(Resource):
     def post(self):
@@ -172,133 +167,6 @@ class LTIStatusAPI(Resource):
         return { "status" : status }
 
 api.add_resource(LTIStatusAPI, '/status')
-
-# /course/:course_uuid/link
-class LTICourseLinkAPI(Resource):
-    @login_required
-    def post(self, course_uuid):
-        """
-        link current session's lti context with a course
-        """
-        course = Course.get_active_by_uuid_or_404(course_uuid)
-        require(EDIT, course,
-            title="Course Not Linked",
-            message="Sorry, you do not have permission to link this course since you are not enrolled as an instructor in the course.")
-
-        if not sess.get('LTI'):
-            abort(400, title="Course Not Linked",
-                message="Sorry, your LTI session has expired. Please log in via LTI and try linking again.")
-
-        if not sess.get('lti_context'):
-            abort(400, title="Course Not Linked",
-                message="Sorry, your LTI link settings have no course context. Please edit your LTI link settings and try linking again.")
-
-        lti_context = LTIContext.query.get_or_404(sess.get('lti_context'))
-        lti_context.compair_course_id = course.id
-        db.session.commit()
-
-        # automatically fetch membership if enabled for context
-        if lti_context.membership_enabled:
-            update_lti_course_membership.delay(course.id)
-
-        on_lti_course_link.send(
-            self,
-            event_name=on_lti_course_link.name,
-            user=current_user,
-            data={ 'course_id': course.id, 'lti_context_id': lti_context.id })
-
-        return { 'success': True }
-
-api.add_resource(LTICourseLinkAPI, '/course/<course_uuid>/link')
-
-# /course/:course_uuid/membership
-class LTICourseMembershipAPI(Resource):
-    @login_required
-    def post(self, course_uuid):
-        """
-        refresh the course membership if able
-        """
-        course = Course.get_active_by_uuid_or_404(course_uuid)
-        require(EDIT, course,
-            title="Membership Not Updated",
-            message="Sorry, your role in this course does not allow you to update membership.")
-
-        if not course.lti_linked:
-            abort(400, title="Membership Not Updated",
-                message="Sorry, your LTI link settings have no course context. Please edit your LTI link settings and try linking again.")
-
-        try:
-            LTIMembership.update_membership_for_course(course)
-        except MembershipNoValidContextsException as err:
-            abort(400, title="Membership Not Updated",
-                message="The LTI link does not support the membership extension. Please edit your LTI link settings or contact your system administrator and try again.")
-        except MembershipNoResultsException as err:
-            abort(400, title="Membership Not Updated",
-                message="The membership service did not return any users. Please check your LTI course and try again.")
-        except MembershipInvalidRequestException as err:
-            abort(400, title="Membership Not Updated",
-                message="The membership request was invalid. Please relaunch the LTI link and try again.")
-
-        on_lti_course_membership_update.send(
-            self,
-            event_name=on_lti_course_membership_update.name,
-            user=current_user,
-            data={ 'course_id': course.id })
-
-        return { 'imported': True }
-
-api.add_resource(LTICourseMembershipAPI, '/course/<course_uuid>/membership')
-
-
-# /course/:course_uuid/membership/status
-class LTICourseMembershipStatusAPI(Resource):
-    @login_required
-    def get(self, course_uuid):
-        """
-        refresh the course membership if able
-        """
-        course = Course.get_active_by_uuid_or_404(course_uuid)
-        require(EDIT, course,
-            title="Membership Status Unavailable",
-            message="Sorry, your role in this course does not allow you to view LTI membership status.")
-
-        if not course.lti_linked:
-            abort(400, title="Membership Status Unavailable",
-                message="The course is not linked to an LTI context yet. Launch an LTI link to link this course first, then check the status.")
-
-        valid_membership_contexts = [
-            lti_context for lti_context in course.lti_contexts \
-            if lti_context.membership_enabled
-        ]
-
-        pending = 0
-        enabled = len(valid_membership_contexts) > 0
-        if enabled:
-            lti_context_ids = [lti_context.id for lti_context in valid_membership_contexts]
-
-            pending = LTIMembership.query \
-                .join(LTIUser) \
-                .filter(and_(
-                    LTIUser.compair_user_id == None,
-                    LTIMembership.lti_context_id.in_(lti_context_ids)
-                )) \
-                .count()
-
-        status = {
-            'enabled': enabled,
-            'pending': pending
-        }
-
-        on_lti_course_membership_status_get.send(
-            self,
-            event_name=on_lti_course_membership_status_get.name,
-            user=current_user,
-            data={ 'course_id': course.id })
-
-        return { 'status': status }
-
-api.add_resource(LTICourseMembershipStatusAPI, '/course/<course_uuid>/membership/status')
-
 
 class ComPAIRRequestValidator(RequestValidator):
     @property
