+++ compair/models/third_party_user.py
@@ -17,7 +17,7 @@ from compair.core import db
 
 # Flask-Login requires the user class to have some methods, the easiest way
 # to get those methods is to inherit from the UserMixin class.
-class ThirdPartyUser(DefaultTableMixin, WriteTrackingMixin):
+class ThirdPartyUser(DefaultTableMixin, UUIDMixin, WriteTrackingMixin):
     __tablename__ = 'third_party_user'
 
     # table columns
