+++ compair/static/less/compair.less
@@ -30,6 +30,9 @@
 // EVALUATE SCREEN
 @import 'comparison.less';
 
+// LTI MANAGE SCREEN
+@import 'lti_manage.less';
+
 // PDF DIRECTIVES
 @import 'rich-content.less';
