+++ compair/static/less/lti_manage.less
@@ -1,0 +1,9 @@
+
+/***** USER COURSES SCREEN *****/
+
+.lti-contexts-screen {
+
+    .search-contexts {
+        margin-top: 1.7em;
+    }
+}//closes user-courses-screen
