+++ compair/static/modules/course/course-assignments-partial.html
@@ -24,6 +24,13 @@
                     Edit Course
                 </a>
             </span>
+            <span ng-if="canManageUsers && course.lti_linked">
+                <a ng-href="#/course/{{course.id}}/lti/context"
+                    class="btn btn-primary" id="add-assignment-btn">
+                    <i class="fa fa-cogs"></i>
+                    Manage LTI Links
+                </a>
+            </span>
         </div>
     </div>
