+++ compair/static/test/features/view_lti_consumers.feature
@@ -12,6 +12,12 @@ Feature: Manage LTI Consumers
       | consumer_key_1      |
       | consumer_key_2      |
       | consumer_key_3      |
+    And I should see '3' contexts listed
+    And I should see contexts with titles:
+      | context_title         |
+      | Canvas Course One     |
+      | Canvas Course Two     |
+      | Blackboard Course One |
 
   Scenario: Disable LTI consumer as admin
     Given I'm a System Administrator
@@ -24,3 +30,13 @@ Feature: Manage LTI Consumers
     And I'm on 'manage lti' page
     When I set the third consumer's active status to 'Active'
     Then I should see a success message
+
+  Scenario: Unlink LTI context as admin
+    Given I'm a System Administrator
+    And I'm on 'manage lti' page
+    When I unlink the second lti context
+    Then I should see '2' contexts listed
+    And I should see contexts with titles:
+      | context_title         |
+      | Canvas Course One     |
+      | Blackboard Course One |
