+++ compair/tasks/lti_membership.py
@@ -2,6 +2,7 @@ import requests
 
 from compair.core import celery, db
 from compair.models import Course, LTIMembership
+from compair.models.lti_models import MembershipNoValidContextsException
 from flask import current_app
 
 @celery.task(bind=True, autoretry_for=(Exception,),
@@ -11,7 +12,11 @@ def update_lti_course_membership(self, course_id):
     if course:
         current_app.logger.info("Begin LTI Membership update for course with id: "+str(course_id)+" named: "+course.name)
 
-        LTIMembership.update_membership_for_course(course)
+        # allow MembershipNoValidContextsException exceptions to occur without retrying job
+        try:
+            LTIMembership.update_membership_for_course(course)
+        except MembershipNoValidContextsException as err:
+            current_app.logger.warning("Error for LTI Membership update for course with id: "+str(course_id)+" named: "+course.name+". No valid lti contexts are linked to the course")
 
         current_app.logger.info("Compelted LTI Membership update for course with id: "+str(course_id)+" named: "+course.name)
     else:
