+++ compair/templates/index-dev.html
@@ -132,6 +132,7 @@
         <script src="{{ url_for('static', filename='modules/login/login-directive.js') }}"></script>
         <script src="{{ url_for('static', filename='modules/lti/lti-module.js') }}"></script>
         <script src="{{ url_for('static', filename='modules/lti_consumer/lti-consumer-module.js') }}"></script>
+        <script src="{{ url_for('static', filename='modules/lti_context/lti-context-module.js') }}"></script>
         <script src="{{ url_for('static', filename='modules/navbar/navbar-module.js') }}"></script>
         <script src="{{ url_for('static', filename='modules/assignment/assignment-module.js') }}"></script>
         <script src="{{ url_for('static', filename='modules/group/group-module.js') }}"></script>
