+++ compair/tests/alembic.ini
@@ -36,7 +36,7 @@ keys = console
 keys = generic
 
 [logger_root]
-level = WARN
+level = ERROR
 handlers = console
 qualname =
